Adding base to checklist test
Create a signup flow for free blog site
Added test spec that runs until checklist page is hit. More steps to come.
Added a function to existing checklist page that will eventually get used by the test. It will use the ID of the task step, which correlates to an update to wp-calypso (using the same branch name)